+++ src/index.ts
@@ -2,3 +2,4 @@ export { generateNumbersArray } from './functions/generate-numbers-array';
 export { uniquifyArray, uniquifyObjectsArray } from './functions/uniquify-array';
 export { objectsArrayToObject } from './functions/objects-array-to-object';
 export { sortArrayByObjectProperty } from './functions/sort-array-by-object-property';
+export { groupArrayByObjectProperty } from './functions/group-array-by-object-property';