+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - uniquifyObjectsArray function.
+- groupArrayByObjectProperty function.
 
 ## [2.0.0] - 2019-11-14
 ### BREAKING CHANGES